feat: Add new SDK route for all-time leader
This commit adds a new SDK route for retrieving the all-time leader in the NBA. The route is implemented in the `server-all-time.py` file and can be accessed via the `/all-time/leader` endpoint. This enhancement expands the functionality of the SDK module and provides users with valuable all-time leaderboard data.