feat: Add franchise player directory path to SDK
This commit adds the franchise player directory path to the SDK module in the `config/path-dir.ts` file. The new directory path is `/franchise/leader/player` and can be accessed via the corresponding route. This enhancement expands the functionality of the SDK module by providing access to franchise player data.