feat: Add route for retrieving synergy data in SDK
This commit adds a new route `/sdk/synergy` in the SDK module for retrieving synergy data. The route is implemented in the `SDKController` class and can be accessed via a GET request. This enhancement expands the functionality of the SDK module by providing access to synergy data based on playtype, team ID, season, and play type nullable parameters.