Add player comparison route in SDK controller
This commit adds a new route `/sdk/player/compare` in the SDK module for retrieving player comparison data. The route is implemented in the `SDKController` class and can be accessed via a GET request. This enhancement expands the functionality of the SDK module by providing access to player comparison data based on various query parameters.