Add GameDate class to handle date-related operations
This commit adds a new file `lib.py` in the `sdk` directory. The file contains a `GameDate` class that provides methods for formatting and retrieving dates. The `get_date_format` method formats a given datetime object to a string representation of the date. The `get_date` method takes in year, month, and day parameters and returns the corresponding date as a string. The `default` attribute is set to the current date. This addition enhances the functionality of the SDK module by providing a convenient way to work with dates.