feat: Add SDKSetController.createPlayerBase route for creating player base
The code changes in the `sdk.ts` file add a new route `PATH_SDK.PLAYER` to the SDK module. This route is responsible for creating a player base using the `SDKSetController.createPlayerBase` method. This addition enhances the functionality of the SDK module by providing a way to create player bases.