SDK for Team Controller #22: Add route to retrieve team roster
This commit adds a new route '/team/{id}/roster' to the SDK server. The route allows users to retrieve the roster of a specific NBA team by providing the team ID as a parameter. The roster data is returned in JSON format. This addition enhances the functionality of the SDK by providing access to team roster information.