Create team switch-case logic | sdk controller #23: Add 'TEAM_ABBV_CALL' directory path to PATH_SDK
This commit adds the 'TEAM_ABBV_CALL' directory path to the PATH_SDK object in the 'path-dir.ts' file. The 'TEAM_ABBV_CALL' property represents the route '/team/abbv/:abbv' and allows for retrieving the team roster directory path for a specific team abbreviation. This addition enhances the functionality of the SDK by providing a convenient way to access team roster information using the team abbreviation as a parameter.