[Initialize] Player Controller #29: Add TEAM_SEASON_STATS directory path to SDK_DIR
This commit adds the TEAM_SEASON_STATS directory path to the SDK_DIR object in the sdk-dir.ts file. The TEAM_SEASON_STATS property represents the route '/team/season/{abbv}' and allows for retrieving the team season stats directory path for a specific team abbreviation. This addition enhances the functionality of the SDK by providing a convenient way to access team season stats information using the team abbreviation as a parameter.