Create Team Base #32: Add createTeamBase method to SDKSetController
This commit adds the createTeamBase method to the SDKSetController class in the sdk-set.ts file. The createTeamBase method is responsible for creating team base data by retrieving team information from the NBA API and inserting it into the database. This addition enhances the functionality of the SDK by providing a convenient way to populate the database with team data.