Team: YearlyStats #30: Add getTeamYearlyStatsByAbbv method to SDKController
This commit adds the getTeamYearlyStatsByAbbv method to the SDKController class in the sdk.ts file. The getTeamYearlyStatsByAbbv method is responsible for retrieving the team yearly stats by abbreviation. It handles the route '/sdk/team/:abbv/yearly/stats' and returns the team stats data for the specified abbreviation. This addition enhances the functionality of the SDK by providing a convenient way to access team yearly stats information.