Stats Module #1: Add new stats constants for player and team stats
This commit adds new constants for player and team stats to the tag.ts file. The new constants include STATS, PLAYER_STATS, TEAM_STATS, and GAME_STATS. These constants will be used to represent different types of stats in the application. This addition enhances the functionality of the application by providing a standardized way to handle and display player and team stats.