feat: Add logo and roster properties to Team model
The `logo` property has been added to the `Team` model in the `Team.ts` file. This change allows for the inclusion of the team's logo in the application. Additionally, the `roster` and `rosterHistory` properties have been added to the `Team` model to store references to the team's roster and roster history. This change improves the organization of the code and allows for better management of team information.