feat: Enable gunicorn for improved server performance
The sdk.sh file has been updated to enable the use of gunicorn for improved server performance. This change adds the gunicorn command to start the SDK server with 4 worker processes. By using gunicorn, the application will be able to handle more concurrent requests and provide a better user experience.